+++ Src/platform/stm32g0b1/CMakeLists.txt
@@ -15,6 +15,7 @@ add_executable(${EXECUTABLE}
     ${ROOT_DIR}/Src/common/module.cpp
 
     ${ROOT_DIR}/Src/periphery/adc/circuit_periphery.cpp
+    ${ROOT_DIR}/Src/platform/stm32/pwm/pwm.cpp
     ${PLATFORM_DIR}/stm32f103/adc.cpp
     ${PLATFORM_DIR}/stm32g0b1/pwm.cpp
     ${PLATFORM_DIR}/stm32f103/iwdg.cpp
